+++ CHANGELOG.md
@@ -7,8 +7,15 @@
 - Experimental support for resolving variables.
   The feature is disabled by default since the functionality is rather limited for now.
   Feel free to comment your feedback at [issue #87](https://github.com/NixOS/nix-idea/issues/87).
+- Support for simple spell checking
+- Automatic insertion of closing quotes
+- Support for *Code | Move Element Left/Right* (<kbd>Ctrl</kbd>+<kbd>Alt</kbd>+<kbd>Shift</kbd>+<kbd>←/→</kbd>)
+- Support for *Code | Move Statement Up/Down* (<kbd>Ctrl</kbd>+<kbd>Shift</kbd>+<kbd>↑/↓</kbd>)
 
 ### Changed
+
+- *Code | Move Statement Up/Down* (<kbd>Ctrl</kbd>+<kbd>Shift</kbd>+<kbd>↑/↓</kbd>)
+  no-longer moves individual lines if that would break up expressions.
 
 ### Deprecated